Refactor #88 - Integrate `ActionDispatcher` into `Store`
